fix: Stack gap spacing when dividers preset to match original implementation
