+++ .changeset/whole-taxes-knock.md
@@ -1,0 +1,5 @@
+---
+'@autoguru/overdrive': patch
+---
+
+**Stack:** spacing with dividers calculated to match original implementation
